Update 190311_jsExercises.html
Added the alphabetPosition() function. I solved this using a nested for loop. The first loop, goes through the text input, the 2nd goes through the alphabet array. When there's a match between the current indices that index + 1 is pushed to the result array. Result array is then returned and joined back into a string